extract color
