+++ rust/src/hexchess/mod.rs
@@ -1,4 +1,5 @@
 pub mod bitboard;
+pub mod color;
 pub mod game;
 pub mod hexchess;
 pub mod piece;
