+++ rust/src/hexchess/pieces/king.rs
@@ -1,5 +1,6 @@
-use crate::constants::{Color, HEXBOARD_GRAPH};
 use crate::color;
+use crate::constants::HEXBOARD_GRAPH;
+use crate::hexchess::color::Color;
 use crate::hexchess::hexchess::Hexchess;
 use crate::hexchess::san::San;
 use smallvec::SmallVec;
