+++ rust/src/lib.rs
@@ -1,8 +1,6 @@
 pub mod constants;
 pub mod hexchess;
 pub mod macros;
-
-pub use constants::Color;
 pub use hexchess::hexchess::Hexchess;
 pub use hexchess::piece::Piece;
 pub use hexchess::san::San;
